Move XYTextSimpleAnnotation to suitable package.
